# api/openapi.yaml
components:
  schemas:
...
    ApiKey:
      additionalProperties: false
      properties:
        last4:
          description: "The key's last four characters, to tell it apart: the key itself is never returned."
          type: string
        refused_at:
          description: "When Volvo refused it (regenerated, or its application deleted on the portal): the vehicles are not read until another key. null: not refused."
          format: date-time
          pattern: Z$
          type:
            - string
            - "null"
        set_at:
          description: When it was given.
          format: date-time
          pattern: Z$
          type: string
      required:
        - last4
        - set_at
        - refused_at
      type: object
...
    Bounds:
      additionalProperties: false
      properties:
        after:
          format: date-time
          pattern: Z$
          type: string
        before:
          format: date-time
          pattern: Z$
          type: string
      required:
        - after
        - before
      type: object
...
    Charge:
      additionalProperties: false
      properties:
        address:
          description: "The street and town of position, from the instance's reverse geocoder, outside every place. null at a place, without a geocoder, when it knows none, and until it answered: a request made by this read is resolved for a later one."
          type:
            - string
            - "null"
        capacity_kwh:
          description: "The battery capacity the energy estimate rests on, as of the readings that ended the event. null when none was known, and for an event derived before Runsten recorded it: a rebuild of the derived data gives it."
          exclusiveMinimum: 0
          format: double
          type:
            - number
            - "null"
        capacity_source:
          description: "Where capacity_kwh comes from. catalog_net: the net capacity of the vehicle's variant in Runsten's catalog; an assumption, the state of charge running from 0 to 100 % over the usable capacity. api: the capacity the vendor reports, gross as far as the published values tell, when the catalog does not know the variant or its net capacity. null with capacity_kwh."
          oneOf:
            - enum:
                - catalog_net
                - api
              type: string
            - type: "null"
        cost:
          description: "What the charge cost: entered by the user, or else estimated from the tariff of its place. null when unknown: without the account's currency, outside every place without an entered cost, without energy_soc_kwh, or when the tariff has no version for a part of the charge's window."
          oneOf:
            - $ref: "#/components/schemas/Cost"
            - type: "null"
        detected_at:
          description: The first reading that revealed the event.
          format: date-time
          pattern: Z$
          type: string
        end:
          $ref: "#/components/schemas/Bounds"
          description: When the event ended.
        end_soc_pct:
          format: double
          type:
            - number
            - "null"
        energy_power_kwh:
          description: "The other estimate: the charging power integrated over the observed readings. null without observed power readings, as for a reconstructed charge."
          format: double
          type:
            - number
            - "null"
        energy_soc_kwh:
          description: "One of two separate estimates, neither a measurement (the vehicle exposes no energy meter): ΔSoC × capacity_kwh."
          format: double
          type:
            - number
            - "null"
        id:
          description: "Identifies the event for its vehicle: its detection time, to the microsecond. Unlike a row ID, it survives a rebuild of the derived data, as long as the same reading reveals the event. Treat it as opaque."
          examples:
            - "2026-09-28T07:01:00.000000Z"
          type: string
        place:
          description: "The place of the charge: the nearest one whose circle holds its position; without a position, the place that takes such charges, unless the charge is DC. null elsewhere."
          oneOf:
            - $ref: "#/components/schemas/PlaceRef"
            - type: "null"
        position:
          oneOf:
            - $ref: "#/components/schemas/Position"
            - type: "null"
        reconstructed:
          description: "The event was never seen, only revealed by the odometer or the state of charge: it lies somewhere within [start.after, end.before], and start and end are that same interval."
          type: boolean
        start:
          $ref: "#/components/schemas/Bounds"
          description: When the event started. The bounds are never collapsed into a single time.
        start_soc_pct:
          format: double
          type:
            - number
            - "null"
        target_soc_pct:
          format: double
          type:
            - number
            - "null"
        type:
          oneOf:
            - enum:
                - AC
                - DC
              type: string
            - type: "null"
      required:
        - type
        - start_soc_pct
        - end_soc_pct
        - target_soc_pct
        - energy_soc_kwh
        - energy_power_kwh
        - position
        - place
        - address
        - cost
        - id
        - detected_at
        - reconstructed
        - start
        - end
        - capacity_kwh
        - capacity_source
      type: object
...
    ConnectionSettings:
      additionalProperties: false
      properties:
        api_key:
          description: "The account's own application key; null: none. Not used while instance_key is true."
          oneOf:
            - $ref: "#/components/schemas/ApiKey"
            - type: "null"
        client_id:
          description: "The client ID of the instance's Volvo application (RUNSTEN_VOLVO_CLIENT_ID): the Volvo IDs connect through it, and its secret, never returned, refreshes the tokens."
          type: string
        connected:
          description: A Volvo ID is connected to the account.
          type: boolean
        instance_key:
          description: "The instance's own application key, which reads every vehicle: an account may not give one. null: none, the account's key is required, before connecting a Volvo ID."
          oneOf:
            - $ref: "#/components/schemas/InstanceKey"
            - type: "null"
      required:
        - connected
        - client_id
        - instance_key
        - api_key
      type: object
    Cost:
      additionalProperties: false
      properties:
        currency:
          description: The account's currency, the unit of min_minor and max_minor.
          enum:
            - CHF
            - CZK
            - DKK
            - EUR
            - GBP
            - ISK
            - NOK
            - PLN
            - SEK
            - USD
          type: string
        efficiency:
          description: "For a tariff, the assumed share of the energy drawn from the grid that reaches the battery: the place's, or else the default of the settings (default_efficiency: ac for AC and unknown types, dc for DC). null for an entered cost."
          format: double
          type:
            - number
            - "null"
        energy_kwh:
          description: "The billed energy. For a tariff, an estimate: energy_soc_kwh (itself an estimate) over efficiency. For an entered cost, the energy entered with it, as read on the receipt; null if none was."
          format: double
          type:
            - number
            - "null"
        max_minor:
          description: The highest amount the charge may have cost.
          format: int64
          minimum: 0
          type: integer
        min_minor:
          description: "The lowest amount the charge may have cost, in minor units of the currency (cents): the energy went through somewhere within [start.after, end.before], at prices that may change over it, so every amount the bounds allow lies within [min_minor, max_minor]. Equal to max_minor when the price is the same over the whole window, and for an entered cost."
          format: int64
          minimum: 0
          type: integer
        note:
          description: The note entered with the cost; null for a tariff, and without a note.
          type:
            - string
            - "null"
        source:
          description: "tariff: estimated from the tariff of the charge's place. entered: the amount the user entered for the charge, which replaces the tariff's."
          enum:
            - tariff
            - entered
          type: string
      required:
        - currency
        - min_minor
        - max_minor
        - source
        - energy_kwh
        - efficiency
        - note
      type: object
...
    Error:
      additionalProperties: false
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
      required:
        - error
      type: object
    ErrorDetail:
      additionalProperties: false
      properties:
        code:
          description: Stable and machine-readable.
          enum:
            - invalid_parameter
            - invalid_body
            - unauthorized
            - invalid_credentials
            - cross_origin
            - not_found
            - method_not_allowed
            - unsupported_media_type
            - too_many_attempts
            - currency_in_use
            - without_position_taken
            - too_many_places
            - charge_has_cost
            - api_key_refused
            - instance_key
            - too_many_vehicles
            - feature_unavailable
            - broker_refused
            - insufficient_scope
            - rate_limited
            - internal
          type: string
        message:
          description: For humans; it may change.
          minLength: 1
          type: string
      required:
        - code
        - message
      type: object
...
    InstanceKey:
      additionalProperties: false
      properties:
        last4:
          description: "The last four characters of the instance's key (RUNSTEN_VOLVO_API_KEY), to check it is the one of the application: the key itself is never returned."
          type: string
      required:
        - last4
      type: object
...
    PageCharge:
      additionalProperties: false
      properties:
        items:
          items:
            $ref: "#/components/schemas/Charge"
          type: array
        next_cursor:
          description: The cursor of the next page; null on the last one.
          type:
            - string
            - "null"
      required:
        - items
        - next_cursor
      type: object
    PageTrip:
      additionalProperties: false
      properties:
        items:
          items:
            $ref: "#/components/schemas/Trip"
          type: array
        next_cursor:
          description: The cursor of the next page; null on the last one.
          type:
            - string
            - "null"
      required:
        - items
        - next_cursor
      type: object
...
    PlaceRef:
      additionalProperties: false
      properties:
        id:
          format: uuid
          type: string
        name:
          type: string
      required:
        - id
        - name
      type: object
...
  securitySchemes:
    accessToken:
      description: "A personal access token, `rst_` then 43 characters, issued by `POST /tokens`: it reads only."
      scheme: bearer
      type: http
    session:
      description: The session cookie set by `POST /session`, when the instance is served over plain http.
      in: cookie
      name: runsten_session
      type: apiKey
    sessionSecure:
      description: "The same cookie when the instance is served over https (its redirect URI is https): `Secure`, and the `__Host-` prefix binds it to the host."
      in: cookie
      name: __Host-runsten_session
      type: apiKey
info:
  contact:
    name: Runsten
    url: https://github.com/runsten-app/runsten
  description: |-
    The session, the account's vehicles, their current state (Verdandi), their
    trips and charges (Urd) with the cost of each charge, their statistics, the account's
    settings (its currency, and its places with their tariffs), and the costs entered for
    charges.

    Conventions:

    - Units are in the field names: `_km`, `_m` (metres), `_kwh`, `_pct` (state of charge, in %), `_w`, `_kw`, `_kwh_per_100km`, `_s` (seconds), `_minor` (minor units of the account's currency, as cents: an integer), `price_per_kwh` (major units of the account's currency per kWh).
      Positions are WGS 84, in degrees.
    - Times are RFC 3339, in UTC.
    - An unknown value is `null`, never zero. Every key is always present.
    - Lists are `{"items": [...]}`, never `null`. Trips and charges are paged with an
      opaque cursor, newest first.
    - Errors are `{"error": {"code", "message"}}`: the code is stable, the message is
      for humans and may change. A path under `/api/` that serves no route answers 404
      `not_found`, and a method that a path does not accept answers 405
      `method_not_allowed` with an `Allow` header.
    - Every response carries `Cache-Control: no-store`, `X-Content-Type-Options: nosniff`
      and `Referrer-Policy: no-referrer`.

    Every route but `POST /session` requires the session cookie it sets, or, to read, a
...
  license:
...
    name: AGPL-3.0-or-later
...
openapi: 3.1.0
paths:
...
  /connection:
    get:
      description: Whether a Volvo ID is connected, and the application key its vehicles are read with.
      operationId: getConnection
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionSettings"
          description: The connection.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`unauthorized`: no valid session, or an invalid or expired access token (with a `WWW-Authenticate` header)."
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`rate_limited`: too many requests with this access token."
          headers:
            Retry-After:
              description: Seconds to wait before the next request.
              required: true
              schema:
                minimum: 1
                type: integer
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`internal`: an internal failure, without detail; the cause is logged."
      security:
        - session: []
        - sessionSecure: []
        - accessToken: []
      summary: The account's connection to Volvo
      tags:
        - connection
...
  /vehicles/{vehicle}/charges:
    get:
      description: Newest first, by `start.after` then `detected_at`. Figures are `null` when the readings do not give them.
      operationId: listCharges
      parameters:
        - description: The vehicle ID.
          in: path
          name: vehicle
          required: true
          schema:
            description: The vehicle ID.
            type: string
        - description: "Keeps the events that may have happened, at least partly, at or after this time: those with end.before after it. An event whose bounds straddle it is kept."
          explode: false
          in: query
          name: from
          schema:
            description: "Keeps the events that may have happened, at least partly, at or after this time: those with end.before after it. An event whose bounds straddle it is kept."
            examples:
              - "2026-09-28T00:00:00Z"
            format: date-time
            type: string
        - description: "Keeps the events that may have happened, at least partly, before this time: those with start.after before it. With from, it must come after from."
          explode: false
          in: query
          name: to
          schema:
            description: "Keeps the events that may have happened, at least partly, before this time: those with start.after before it. With from, it must come after from."
            examples:
              - "2026-09-29T00:00:00Z"
            format: date-time
            type: string
        - description: The next_cursor of the previous page. It stays valid when from or to change.
          explode: false
          in: query
          name: cursor
          schema:
            description: The next_cursor of the previous page. It stays valid when from or to change.
            type: string
        - description: The maximum number of events of the page.
          explode: false
          in: query
          name: limit
          schema:
            default: 50
            description: The maximum number of events of the page.
            format: int64
            maximum: 200
            minimum: 1
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageCharge"
          description: A page of charges, newest first.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`invalid_parameter`: `from`, `to`, `limit` or `cursor`; the message names the parameter."
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`unauthorized`: no valid session, or an invalid or expired access token (with a `WWW-Authenticate` header)."
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`not_found`: no such vehicle in the account. Another account's vehicle is not found either."
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`rate_limited`: too many requests with this access token."
          headers:
            Retry-After:
              description: Seconds to wait before the next request.
              required: true
              schema:
                minimum: 1
                type: integer
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`internal`: an internal failure, without detail; the cause is logged."
      security:
        - session: []
        - sessionSecure: []
        - accessToken: []
      summary: The charges of a vehicle (Urd)
      tags:
        - events
...
  /vehicles/{vehicle}/trips:
    get:
      description: Newest first, by `start.after` then `detected_at`. Figures are `null` when the readings do not give them.
      operationId: listTrips
      parameters:
        - description: The vehicle ID.
          in: path
          name: vehicle
          required: true
          schema:
            description: The vehicle ID.
            type: string
        - description: "Keeps the events that may have happened, at least partly, at or after this time: those with end.before after it. An event whose bounds straddle it is kept."
          explode: false
          in: query
          name: from
          schema:
            description: "Keeps the events that may have happened, at least partly, at or after this time: those with end.before after it. An event whose bounds straddle it is kept."
            examples:
              - "2026-09-28T00:00:00Z"
            format: date-time
            type: string
        - description: "Keeps the events that may have happened, at least partly, before this time: those with start.after before it. With from, it must come after from."
          explode: false
          in: query
          name: to
          schema:
            description: "Keeps the events that may have happened, at least partly, before this time: those with start.after before it. With from, it must come after from."
            examples:
              - "2026-09-29T00:00:00Z"
            format: date-time
            type: string
        - description: The next_cursor of the previous page. It stays valid when from or to change.
          explode: false
          in: query
          name: cursor
          schema:
            description: The next_cursor of the previous page. It stays valid when from or to change.
            type: string
        - description: The maximum number of events of the page.
          explode: false
          in: query
          name: limit
          schema:
            default: 50
            description: The maximum number of events of the page.
            format: int64
            maximum: 200
            minimum: 1
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageTrip"
          description: A page of trips, newest first.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`invalid_parameter`: `from`, `to`, `limit` or `cursor`; the message names the parameter."
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`unauthorized`: no valid session, or an invalid or expired access token (with a `WWW-Authenticate` header)."
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`not_found`: no such vehicle in the account. Another account's vehicle is not found either."
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`rate_limited`: too many requests with this access token."
          headers:
            Retry-After:
              description: Seconds to wait before the next request.
              required: true
              schema:
                minimum: 1
                type: integer
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`internal`: an internal failure, without detail; the cause is logged."
      security:
        - session: []
        - sessionSecure: []
        - accessToken: []
      summary: The trips of a vehicle (Urd)
      tags:
        - events
...
servers:
...